fix(group): use new conversation endpoint
The old groups endpoint were deprecated and stopped working on Feb 24.
Replace any instances of the old API with the new conversations API.
Unfortunately, the new API doesn't return a list of members for each channel,
so we need to make an extra call for each channel to fetch the list of
members (required to filter out channels the user are already in and to show
the number of members in each channel).

Ref: https://api.slack.com/changelog/2020-01-deprecating-antecedents-to-the-conversations-api